diagnostics/cnn_service.py
# diagnostics/cnn_service.py
# Real CNN preprocessing pipeline for cardiac CT scans
# FR-08: AI CT Evaluation — Preprocessing Pipeline
# Reference: .agent/ohms.md — Section 8

import os
import numpy as np
from PIL import Image, ImageOps, ImageFilter
import pydicom
import cv2
import logging
from django.conf import settings

logger = logging.getLogger(__name__)

# ---------------------------------------------------------------------------
# Global model reference — loaded once at startup in diagnostics/apps.py
# ---------------------------------------------------------------------------
_model = None

def load_cnn_model(model_path: str) -> None:
    """
    Load the CNN model from disk into the global _model variable.
    Called once in DiagnosticsConfig.ready().
    """
    global _model
    try:
        import tensorflow as tf
        _model = tf.keras.models.load_model(model_path)
        logger.info(f"[CNN Service] Successfully loaded model from {model_path}")
    except Exception as e:
        logger.warning(f"[CNN Service] Failed to load real model: {str(e)}. Falling back to stub.")
        _model = "STUB_MODEL"

def preprocess(image_path: str) -> str:
    """
    Real Preprocessing Pipeline (Step 8 Implementation).
    
    Order:
      1. Open image (Regular or DICOM)
      2. Convert to grayscale
      3. Resize 224x224
      4. Median filter (Noise reduction)
      5. Contrast enhancement (Equalize)
      6. Normalize [0, 1] (Internal step)
      7. Save as PNG to /media/ct_scans/preprocessed/
    """
    try:
        logger.info(f"[CNN Service] Starting preprocessing for: {image_path}")
        
        # 1. Open image file
        ext = os.path.splitext(image_path)[1].lower()
        if ext == '.dcm':
            ds = pydicom.dcmread(image_path)
            # Convert DICOM pixel array to PIL Image
            # Handle potential windowing/scaling in DICOM
            img_array = ds.pixel_array.astype(float)
            # Rescale to 0-255
            img_array = (np.maximum(img_array, 0) / img_array.max()) * 255.0
            img = Image.fromarray(img_array.astype(np.uint8))
        else:
            img = Image.open(image_path)
        
        # 2. Convert to grayscale
        img = img.convert('L')
        
        # 3. Resize to 224 x 224
        img = img.resize((224, 224), Image.Resampling.LANCZOS)
        
        # 4. Median filter for noise reduction
        img = img.filter(ImageFilter.MedianFilter(size=3))
        
        # 5. Contrast enhancement (PIL Equalize)
        img = ImageOps.equalize(img)
        
        # 6. Normalize pixel values to [0, 1] range (Internal representation check)
        # For the final PNG save, we keep it as 0-255, but we've ensured the range is full.
        img_array = np.array(img).astype(float) / 255.0
        # Pipeline metadata verification:
        logger.debug(f"[CNN Service] Normalization check: min={img_array.min()}, max={img_array.max()}")
        
        # Convert back to uint8 for saving
        final_img = Image.fromarray((img_array * 255.0).astype(np.uint8))
        
        # 7. Save preprocessed image as PNG
        original_filename = os.path.basename(image_path)
        filename_no_ext = os.path.splitext(original_filename)[0]
        preprocessed_filename = f"preprocessed_{filename_no_ext}.png"
        
        save_dir = os.path.join(settings.MEDIA_ROOT, 'ct_scans', 'preprocessed')
        if not os.path.exists(save_dir):
            os.makedirs(save_dir)
            
        save_path = os.path.join(save_dir, preprocessed_filename)
        final_img.save(save_path, 'PNG')
        
        # 8. Return relative path for storage
        relative_path = os.path.join('ct_scans', 'preprocessed', preprocessed_filename)
        logger.info(f"[CNN Service] Preprocessing complete: {relative_path}")
        return relative_path

    except Exception as e:
        logger.error(f"[CNN Service] Preprocessing FAILED for {image_path}: {str(e)}")
        # Caller (view) handles status update to 'failed' based on None return
        return None

def predict(preprocessed_path: str) -> dict:
    """
    Real CNN inference (Phase A Implementation).
    
    Inference code:
      - Load preprocessed image
      - Convert to grayscale, resize 224x224, normalize [0,1]
      - Reshape to (1, 224, 224, 1)
      - Run model.predict()
    """
    global _model
    
    # 1. Graceful fallback if model not loaded
    if _model is None or _model == "STUB_MODEL":
        logger.debug(f"[CNN STUB] predict() falling back to stub for: {preprocessed_path}")
        return {
            "class": "normal",
            "confidence": 0.88,
            "model_version": "stub-v1-fallback"
        }

    try:
        # 2. Real Inference Pipeline
        # Open and ensure grayscale
        img = Image.open(preprocessed_path).convert('L')
        # Resize to exactly 224x224
        img = img.resize((224, 224))
        # Convert to numpy and normalize to [0, 1]
        arr = np.array(img).astype('float32') / 255.0
        # Add batch and channel dimensions: (1, 224, 224, 1)
        arr = arr.reshape(1, 224, 224, 1)
        
        # 3. Predict
        prediction = _model.predict(arr, verbose=0)
        score = float(prediction[0][0])
        
        # 4. Result interpretation
        # >= 0.5 -> Abnormal
        is_abnormal = score >= 0.5
        confidence = score if is_abnormal else (1.0 - score)
        
        return {
            "class": "abnormal" if is_abnormal else "normal",
            "confidence": round(confidence, 4),
            "model_version": "cardiac-cnn-v1"
        }

    except Exception as e:
        logger.error(f"[CNN Service] Inference FAILED: {str(e)}")
        return {
            "class": "error",
            "confidence": 0.0,
            "error_msg": str(e)
        }
