requirements.txt
Django==6.0.4
mysqlclient==2.2.8
python-dotenv==1.0.1
Pillow==11.2.1
numpy==2.2.2
scipy==1.15.2
pydicom==3.0.2
opencv-python==4.11.0.86
tensorflow>=2.15.0
